Rename module to os
