+++ .github/workflows/publish-release.yml
@@ -42,13 +42,13 @@ jobs:
                   ./gradlew -Pversion=${VERSION} publish -x test
             - name: Create Github release from the release tag
               run: |
-                  curl --request POST 'https://api.github.com/repos/ballerina-platform/module-ballerina-system/releases' \
+                  curl --request POST 'https://api.github.com/repos/ballerina-platform/module-ballerina-os/releases' \
                   --header 'Accept: application/vnd.github.v3+json' \
                   --header 'Authorization: Bearer ${{ secrets.BALLERINA_BOT_TOKEN }}' \
                   --header 'Content-Type: application/json' \
                   --data-raw '{
                   "tag_name": "v'"$VERSION"'",
-                  "name": "module-ballerina-system-v'"$VERSION"'"
+                  "name": "module-ballerina-os-v'"$VERSION"'"
                   }'
             - name: Post release PR
               env:
