+++ build.gradle
@@ -58,7 +58,7 @@ allprojects {
         }
 
         maven {
-            url = 'https://maven.pkg.github.com/ballerina-platform/module-ballerina-system'
+            url = 'https://maven.pkg.github.com/ballerina-platform/module-ballerina-os'
             credentials {
                 username System.getenv("packageUser")
                 password System.getenv("packagePAT")
@@ -106,7 +106,7 @@ release {
 }
 
 task build {
-    dependsOn('system-ballerina:build')
+    dependsOn('os-ballerina:build')
 }
 
 task codeCoverageReport(type: JacocoReport) {
