+++ os-ballerina/Ballerina.toml
@@ -1,0 +1,30 @@
+[package]
+org = "ballerina"
+name = "os"
+version = "@toml.version@"
+
+[[dependency]]
+org = "ballerina"
+name = "io"
+version = "@stdlib.io.version@"
+
+[[platform.java11.dependency]]
+path = "../os-native/build/libs/os-native-@project.version@.jar"
+groupId = "ballerina"
+artifactId = "os"
+version = "@project.version@"
+
+
+[[platform.java11.dependency]]
+path = "./lib/io-native-@io.native.version@.jar"
+groupId = "org.ballerinalang"
+artifactId = "io-native"
+version = "@io.native.version@"
+
+
+[[platform.java11.dependency]]
+path = "../os-test-utils/build/libs/os-test-utils-@project.version@.jar"
+groupId = "org.ballerinalang.stdlib.os"
+scope = "testOnly"
+artifactId = "os"
+version = "@project.version@"
