+++ os-ballerina/Package.md
@@ -1,0 +1,5 @@
+## Module overview
+
+This module provides functions to retrieve information about the operating system and the current users of the operating system. The `os:exec` method creates the `os:Process` object, which can control and obtain information about a Ballerina process.
+
+For information on the operations, which you can perform with the os module, see the below **Functions**.
