+++ os-ballerina/system_errors.bal
@@ -1,0 +1,33 @@
+// Copyright (c) 2019 WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+//
+// WSO2 Inc. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+# Represents an `InvalidOperationError` with the message and the cause.
+public type InvalidOperationError distinct error;
+
+# Represents a `PermissionError` with the message and the cause.
+public type PermissionError distinct error;
+
+# Represents a `FileSystemError` with the message and the cause.
+public type FileSystemError distinct error;
+
+# Represents a `FileNotFoundError` with the message and the cause.
+public type FileNotFoundError distinct error;
+
+# Represents an `ProcessExecError` with the message and the cause.
+public type ProcessExecError distinct error;
+
+# The union of all the OS related errors.
+public type Error InvalidOperationError|PermissionError|FileSystemError|FileNotFoundError|ProcessExecError;
