--- system-ballerina/Ballerina.toml
@@ -1,30 +1,0 @@
-[package]
-org = "ballerina"
-name = "system"
-version = "@toml.version@"
-
-[[dependency]]
-org = "ballerina"
-name = "io"
-version = "@stdlib.io.version@"
-
-[[platform.java11.dependency]]
-path = "../system-native/build/libs/system-native-@project.version@.jar"
-groupId = "ballerina"
-artifactId = "system"
-version = "@project.version@"
-
-
-[[platform.java11.dependency]]
-path = "./lib/io-native-@io.native.version@.jar"
-groupId = "org.ballerinalang"
-artifactId = "io-native"
-version = "@io.native.version@"
-
-
-[[platform.java11.dependency]]
-path = "../system-test-utils/build/libs/system-test-utils-@project.version@.jar"
-groupId = "org.ballerinalang.stdlib.system"
-scope = "testOnly"
-artifactId = "system"
-version = "@project.version@"
